+++ ChangeLog.md
@@ -4,3 +4,7 @@
 # 0.0.2
 - Forked from `foc`.
 - Added `probify`
+
+# 0.0.3
+- bug fixes: `dmap`, `probify`, and `taskbar`
+- `dmap` supports symoblic merging and updates.